Rename keyboardlayout to keyboard
