+++ keyboard/layout.go
@@ -1,5 +1,5 @@
-// Package keyboardlayout provides keyboard key interpretation and layout validation.
-package keyboardlayout
+// Package keyboard provides keyboard key interpretation and layout validation.
+package keyboard
 
 import (
 	"fmt"
